Improvement: DRY db connection related stuff
