+++ src/Checks/Checks/DatabaseSizeCheck.php
@@ -6,19 +6,13 @@ use Illuminate\Database\ConnectionResolverInterface;
 use Spatie\Health\Checks\Check;
 use Spatie\Health\Checks\Result;
 use Spatie\Health\Support\DbConnectionInfo;
+use Spatie\Health\Traits\DatabaseRelatable;
 
 class DatabaseSizeCheck extends Check
 {
-    protected ?string $connectionName = null;
+    use DatabaseRelatable;
 
     protected float $failWhenSizeAboveGb = 1;
-
-    public function connectionName(string $connectionName): self
-    {
-        $this->connectionName = $connectionName;
-
-        return $this;
-    }
 
     public function failWhenSizeAboveGb(float $errorThresholdGb): self
     {
@@ -40,11 +34,6 @@ class DatabaseSizeCheck extends Check
         return $databaseSizeInGb >= $this->failWhenSizeAboveGb
             ? $result->failed("Database size is {$databaseSizeInGb} GB, which is above the threshold of {$this->failWhenSizeAboveGb} GB")
             : $result->ok();
-    }
-
-    protected function getDefaultConnectionName(): string
-    {
-        return config('database.default');
     }
 
     protected function getDatabaseSizeInGb(): float
